Project Development Phase/Sprint 4/requirements.txt
django
djangorestframework
knox
django-rest-knox
ibm-db
python-dotenv
newsapi-python